+++ lib/simpler_command/railtie.rb
@@ -5,6 +5,7 @@ module SimplerCommand
   class Railtie < ::Rails::Railtie
     generators do
       require "generators/simpler_command/simpler_command_generator"
+      require "generators/rspec/simpler_command/simpler_command_generator" if defined? RSpec::Rails
     end
   end
 end